Add a debugging command
Add a config flag and command that can be used for debugging the bot.
The first sub-command for debugging will announce all of the events
found on the deploy page to the attached irc channel. This obviously
noisy output should not be directed at any "real" channel but instead
used by attaching to a temporary testing channel just for debugging.

Change-Id: I5a27d38fa7cea976ce421624185a326c3b7f825a