Add a "now" command
This comes up occasionally on irc as a desired feature, so let's
implement it. The "now" command does a brute force scan of the event
database and outputs a message for every event where start time if now
or before and end time is now or later. If there are no active deploy
windows then the bot will say so along with the time until the next
window starts.

Change-Id: I6960e576ac846b7e66106e556a9c303e09678cbc