Don't keep adding _ to nick indefinitely
on_nicknameinuse() and do_reclaim_nick() will be called in loop when
multiple instances of the bot are running. This can lead to something
like this:

  [05:26] jouncebot_  joined the channel
  [05:26] jouncebot_ is now known as jouncebot__
  [05:26] jouncebot__ is now known as jouncebot___

This happens because conn.get_nickname() returns the current nick of the
bot and not the nick that was attempted. Use the configured nick instead
so that we will stick with `jouncebot_` as long as `jouncebot` is taken.

Bug: T150916
Change-Id: Id0d59dc180c8e394ad860a7390a05667c2649688