Reparse deploy page before announcing an event
Help with the common problem of last minute wiki page additions which
have not been found via the 15 minute reparse timer by reparsing at the
start of each event window. Fetch + parse time takes <10s typically so
this should not be disruptive to the core function (yelling on irc when
its time to do something).

Bug: T243394
Change-Id: I42ec0346feebb18c7051c26d124d04af5fc6ccff