cleanup: commit some changes from libera.chat migration
There have been a small number of local changes in Toolforge since we
started migrating to libera.chat. For a brief time there were 2 copies
of the bot running, one for each IRC network. That duplication was
handled by making a new deployment and config files. We are back to
a single deployment now, but the change of adding in an identifier for
the IRC network attached seems worth keeping around. I hope that
libera.chat is as stable as the prior network was, but really IRC
networks can come and go at a whim so we may as well leave breadcrumbs
for a future network shuffle.

Change-Id: I9f2e6b7a34c1457f0af1d2fbc597ee120228804c